Match is now has a property to return the sentence
When playing with grammarbot for Python I noticed that the sentence was not returned in the match object. This makes it difficult to identify the sentence the match is referring to. So I have updated the object without diving in to the match json object. So I created a shortcut property like the rest of this object.